Update test_notebooks.yml
The failure occurred because the workflow file is using an outdated version (`v1`) of the `actions/cache` action, which no longer has the required "download info". This issue can be resolved by upgrading to a more recent version of the `actions/cache` action.

### Solution
Update the `.github/workflows/test_notebooks.yml` file as follows:

#### Before:
```yaml
- name: pip cache
  uses: actions/cache@v1
  with:
    path: ~/.cache/pip
    key: lint-pip-${{ hashFiles('**/pyproject.toml') }}
    restore-keys: |
      lint-pip-
```

#### After:
```yaml
- name: pip cache
  uses: actions/cache@v3
  with:
    path: ~/.cache/pip
    key: lint-pip-${{ hashFiles('**/pyproject.toml') }}
    restore-keys: |
      lint-pip-
```

### Steps to Implement
1. Replace `actions/cache@v1` with `actions/cache@v3` in the workflow file.
2. Commit and push the changes to the branch associated with the pull request.
3. Trigger the workflow again to verify the fix.

Would you like me to assist further with creating a commit message or additional details?